+++ extensions/omni_httpd/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-02-04
+
+### Added
+
+* `Omnigres-Connecting-IP` header [#714](https://github.com/omnigres/omnigres/pull/714)
+
+### Changed
+
+* omni_httpd.handler is no longer a function but a procedure [#715](https://github.com/omnigres/omnigres/pull/715)
+
+### Fixed
+
+* Response body is no longer sent before committing the transaction [#778](https://github.com/omnigres/omnigres/pull/778)
+
 ## [0.2.10] - 2025-01-23
 
 ### Added
@@ -133,3 +147,5 @@ Initial release following a few months of iterative development.
 [0.2.9]: [https://github.com/omnigres/omnigres/pull/752]
 
 [0.2.10]: [https://github.com/omnigres/omnigres/pull/763]
+
+[0.3.0]: [https://github.com/omnigres/omnigres/pull/713]